backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
huggingface-hub==0.20.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.0
